+++ cmd/config_windows.go
@@ -13,6 +13,7 @@ var (
 	DefaultConfigPath         = filepath.Join(os.Getenv("AppData"), "gmhost", "config.yml")
 	DefaultCacheLocation      = filepath.Join(os.Getenv("AppData"), "gmhost", "cache.db")
 	DefaultQuarantineLocation = filepath.Join(os.Getenv("AppData"), "gmhost", "quarantine")
+	DefaultPluginsLocation    = filepath.Join(os.Getenv("AppData"), "gmhost", "plugins")
 )
 
 func getConfigFile() (config string) {