+++ cmd/scan_linux.go
@@ -6,6 +6,7 @@ package cmd
 import (
 	"context"
 
+	"github.com/glimps-re/host-connector/pkg/report"
 	"github.com/glimps-re/host-connector/pkg/scanner"
 )
 
@@ -13,6 +14,6 @@ func Gui(entryPath string, nbFiles int) context.Context {
 	return nil
 }
 
-func (a *GuiHandleResult) Handle(ctx context.Context, path string, result scanner.SummarizedGMalwareResult, report *scanner.Report) (err error) {
+func (a *GuiHandleResult) Handle(ctx context.Context, path string, result scanner.SummarizedGMalwareResult, report *report.Report) (err error) {
 	return nil
 }